requirements.txt
streamlit==1.18.0
pandas==1.5.3
python-dotenv==0.21.0
boto3==1.26.129
apify-client==1.1.1
streamlit_option_menu==0.3.2
plotly==5.14.1
s3fs==2023.5.0
git+https://github.com/streamlit/files-connection